requirements.txt
streamlit
easyocr
Pillow
pdf2image
pandas
python-docx
google-generativeai